tests: reorganize runtime and storage suites
