+++ Tests/SpeakSwiftlyTests/Runtime/WorkerRuntimeControlSurfaceTests.swift
@@ -3,26 +3,6 @@ import Foundation
 import Testing
 import TextForSpeech
 
-// MARK: - WeakRuntimeBox
-
-private final class WeakRuntimeBox: @unchecked Sendable {
-    weak var value: WorkerRuntime?
-}
-
-// MARK: - BackendLoadRecorder
-
-private actor BackendLoadRecorder {
-    private var backends = [SpeakSwiftly.SpeechBackend]()
-
-    func record(_ backend: SpeakSwiftly.SpeechBackend) {
-        backends.append(backend)
-    }
-
-    func values() -> [SpeakSwiftly.SpeechBackend] {
-        backends
-    }
-}
-
 @Test func `list queue returns active and queued requests without waiting for active playback`() async throws {
     let output = OutputRecorder()
     let playbackDrain = AsyncGate()
@@ -876,1376 +856,6 @@ private actor BackendLoadRecorder {
     await profileGate.open()
 }
 
-@Test func `library helpers submit profile and generated file worker protocol requests`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let createID = await runtime.voices
-        .create(design: "bright-guide",
-                from: "Hello there",
-                vibe: .femme,
-                voice: "Warm and bright",
-                outputPath: nil)
-        .id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == createID
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "bright-guide"
-        }
-    })
-
-    let listID = await runtime.voices.list().id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == listID,
-                $0["ok"] as? Bool == true,
-                let profiles = $0["profiles"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            return profiles.contains { $0["profile_name"] as? String == "bright-guide" }
-        }
-    })
-
-    let renameID = await runtime.voices.rename("bright-guide", to: "clear-guide").id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == renameID
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "clear-guide"
-        }
-    })
-
-    let renamedListID = await runtime.voices.list().id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == renamedListID,
-                $0["ok"] as? Bool == true,
-                let profiles = $0["profiles"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            let names = profiles.compactMap { $0["profile_name"] as? String }
-            return names.contains("clear-guide") && !names.contains("bright-guide")
-        }
-    })
-
-    let rerollID = await runtime.voices.reroll("clear-guide").id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == rerollID
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "clear-guide"
-        }
-    })
-
-    let speakFileID = await runtime.generate
-        .audio(
-            text: "Save this request as an artifact.",
-            with: "clear-guide",
-        )
-        .id
-    let fileArtifactID = "\(speakFileID)-artifact-1"
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == speakFileID,
-                $0["ok"] as? Bool == true,
-                let generatedFile = $0["generated_file"] as? [String: Any]
-            else {
-                return false
-            }
-
-            return generatedFile["artifact_id"] as? String == fileArtifactID
-        }
-    })
-
-    let generatedFileID = await runtime.artifacts.file(id: fileArtifactID).id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == generatedFileID,
-                $0["ok"] as? Bool == true,
-                let generatedFile = $0["generated_file"] as? [String: Any]
-            else {
-                return false
-            }
-
-            return generatedFile["artifact_id"] as? String == fileArtifactID
-        }
-    })
-
-    let generatedFilesID = await runtime.artifacts.files().id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == generatedFilesID,
-                $0["ok"] as? Bool == true,
-                let generatedFiles = $0["generated_files"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            return generatedFiles.contains {
-                $0["artifact_id"] as? String == fileArtifactID
-            }
-        }
-    })
-
-    let generationJobID = await runtime.jobs.job(id: speakFileID).id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == generationJobID,
-                $0["ok"] as? Bool == true,
-                let generationJob = $0["generation_job"] as? [String: Any]
-            else {
-                return false
-            }
-
-            return generationJob["job_id"] as? String == speakFileID
-                && generationJob["job_kind"] as? String == "file"
-                && generationJob["state"] as? String == "completed"
-                && (generationJob["items"] as? [[String: Any]])?.first?["artifact_id"] as? String == fileArtifactID
-        }
-    })
-
-    let generationJobsID = await runtime.jobs.list().id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == generationJobsID,
-                $0["ok"] as? Bool == true,
-                let generationJobs = $0["generation_jobs"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            return generationJobs.contains {
-                $0["job_id"] as? String == speakFileID
-                    && $0["job_kind"] as? String == "file"
-                    && $0["state"] as? String == "completed"
-                    && ($0["items"] as? [[String: Any]])?.first?["artifact_id"] as? String == fileArtifactID
-            }
-        }
-    })
-
-    let removeID = await runtime.voices.delete(named: "clear-guide").id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == removeID
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "clear-guide"
-        }
-    })
-}
-
-@Test func `reroll rebuilds an existing profile in place from its stored inputs`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let store = try makeProfileStore(rootURL: storeRoot)
-    let originalAudio = Data([0x01, 0x02, 0x03, 0x04])
-    _ = try store.createProfile(
-        profileName: "bright-guide",
-        vibe: .femme,
-        modelRepo: ModelFactory.profileModelRepo,
-        voiceDescription: "Warm and bright.",
-        sourceText: "Hello there",
-        sampleRate: 24000,
-        canonicalAudioData: originalAudio,
-    )
-
-    let rerolledSamples: [Float] = [0.7, 0.8, 0.9]
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-        profileModelLoader: {
-            AnySpeechModel(
-                sampleRate: 24000,
-                generate: { _, _, _, _, _, _ in rerolledSamples },
-                generateSamplesStream: { _, _, _, _, _, _, _ in
-                    AsyncThrowingStream { continuation in
-                        continuation.finish()
-                    }
-                },
-            )
-        },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let rerollID = await runtime.voices.reroll("bright-guide").id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == rerollID
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "bright-guide"
-        }
-    })
-
-    let rerolledProfile = try store.loadProfile(named: "bright-guide")
-    #expect(rerolledProfile.manifest.profileName == "bright-guide")
-    #expect(rerolledProfile.manifest.sourceText == "Hello there")
-    #expect(rerolledProfile.manifest.voiceDescription == "Warm and bright.")
-    #expect(try Data(contentsOf: rerolledProfile.referenceAudioURL) == rawTestAudioData(for: rerolledSamples))
-}
-
-@Test func `create profile resolves relative output path against explicit caller working directory`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    let callerWorkingDirectory = makeTempDirectoryURL()
-    defer {
-        try? FileManager.default.removeItem(at: storeRoot)
-        try? FileManager.default.removeItem(at: callerWorkingDirectory)
-    }
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let exportURL = callerWorkingDirectory.appendingPathComponent("exports/voice.wav")
-    await runtime.accept(
-        line: #"{"id":"req-relative-export","op":"create_voice_profile_from_description","profile_name":"bright-guide","text":"Hello there","vibe":"femme","voice_description":"Warm and bright","output_path":"exports/voice.wav","cwd":"\#(callerWorkingDirectory.path)"}"#,
-    )
-
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == "req-relative-export"
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "bright-guide"
-        }
-    })
-    #expect(FileManager.default.fileExists(atPath: exportURL.path))
-}
-
-@Test func `create profile rejects relative output path without explicit caller working directory`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    await runtime.accept(
-        line: #"{"id":"req-relative-export-missing-cwd","op":"create_voice_profile_from_description","profile_name":"bright-guide","text":"Hello there","vibe":"femme","voice_description":"Warm and bright","output_path":"exports/voice.wav"}"#,
-    )
-
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            ($0["id"] as? String) == "req-relative-export-missing-cwd"
-                && ($0["ok"] as? Bool) == false
-                && ($0["code"] as? String) == "invalid_request"
-                && (($0["message"] as? String)?.contains("did not provide 'cwd'") ?? false)
-        }
-    })
-}
-
-@Test func `generate batch acknowledges queue then completes with generated batch metadata`() async throws {
-    let output = OutputRecorder()
-    let playback = PlaybackSpy()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let store = try makeProfileStore(rootURL: storeRoot)
-    _ = try store.createProfile(
-        profileName: "default-femme",
-        modelRepo: "test-model",
-        voiceDescription: "Warm and bright.",
-        sourceText: "Reference transcript",
-        sampleRate: 24000,
-        canonicalAudioData: Data([0x01, 0x02]),
-    )
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: playback,
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let batchID = await runtime.generate
-        .batch(
-            [
-                SpeakSwiftly.BatchItem(text: "First generated file."),
-                SpeakSwiftly.BatchItem(
-                    artifactID: "custom-batch-artifact",
-                    text: "Second generated file.",
-                    textProfileName: "logs",
-                ),
-            ],
-            with: "default-femme",
-        )
-        .id
-
-    #expect(await waitUntil {
-        output.countJSONObjects {
-            $0["id"] as? String == batchID
-                && $0["ok"] as? Bool == true
-        } == 2
-    })
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == batchID,
-                let generationJob = $0["generation_job"] as? [String: Any],
-                let items = generationJob["items"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            return generationJob["job_id"] as? String == batchID
-                && generationJob["job_kind"] as? String == "batch"
-                && generationJob["state"] as? String == "queued"
-                && items.count == 2
-                && items[0]["artifact_id"] as? String == "\(batchID)-artifact-1"
-                && items[1]["artifact_id"] as? String == "custom-batch-artifact"
-        }
-    })
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == batchID
-                && $0["event"] as? String == "started"
-                && $0["op"] as? String == "generate_batch"
-        }
-    })
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == batchID,
-                let generatedBatch = $0["generated_batch"] as? [String: Any],
-                let generationJob = $0["generation_job"] as? [String: Any],
-                let artifacts = generatedBatch["artifacts"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            return generatedBatch["batch_id"] as? String == batchID
-                && generatedBatch["state"] as? String == "completed"
-                && generationJob["job_kind"] as? String == "batch"
-                && generationJob["state"] as? String == "completed"
-                && artifacts.count == 2
-                && artifacts.contains { $0["artifact_id"] as? String == "\(batchID)-artifact-1" }
-                && artifacts.contains { $0["artifact_id"] as? String == "custom-batch-artifact" }
-        }
-    })
-
-    let generatedBatchID = await runtime.artifacts.batch(id: batchID).id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == generatedBatchID,
-                let generatedBatch = $0["generated_batch"] as? [String: Any],
-                let artifacts = generatedBatch["artifacts"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            return generatedBatch["batch_id"] as? String == batchID
-                && artifacts.count == 2
-        }
-    })
-
-    let generatedBatchesID = await runtime.artifacts.batches().id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == generatedBatchesID,
-                let generatedBatches = $0["generated_batches"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            return generatedBatches.contains {
-                $0["batch_id"] as? String == batchID
-                    && $0["state"] as? String == "completed"
-            }
-        }
-    })
-}
-
-@Test func `create clone stores provided transcript without transcription`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-    try FileManager.default.createDirectory(at: storeRoot, withIntermediateDirectories: true)
-
-    let referenceAudioURL = storeRoot.appendingPathComponent("reference.wav")
-    try Data([0x01]).write(to: referenceAudioURL, options: .atomic)
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        loadedCloneAudioSamples: [0.1, 0.2, 0.3],
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let createID = await runtime.voices
-        .create(clone: "ghost-copy",
-                from: referenceAudioURL,
-                transcript: "Provided transcript")
-        .id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == createID
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "ghost-copy"
-        }
-    })
-    #expect(!output.containsJSONObject {
-        $0["id"] as? String == createID
-            && $0["stage"] as? String == "transcribing_clone_audio"
-    })
-
-    let storedProfile = try makeProfileStore(rootURL: storeRoot).loadProfile(named: "ghost-copy")
-    #expect(storedProfile.manifest.sourceText == "Provided transcript")
-    #expect(storedProfile.manifest.voiceDescription == ModelFactory.importedCloneVoiceDescription)
-    #expect(storedProfile.manifest.modelRepo == ModelFactory.importedCloneModelRepo)
-}
-
-@Test func `create clone resolves relative reference audio against explicit caller working directory`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    let callerWorkingDirectory = makeTempDirectoryURL()
-    defer {
-        try? FileManager.default.removeItem(at: storeRoot)
-        try? FileManager.default.removeItem(at: callerWorkingDirectory)
-    }
-    try FileManager.default.createDirectory(at: storeRoot, withIntermediateDirectories: true)
-    try FileManager.default.createDirectory(at: callerWorkingDirectory, withIntermediateDirectories: true)
-
-    let referenceAudioURL = callerWorkingDirectory.appendingPathComponent("reference.wav")
-    try Data([0x01]).write(to: referenceAudioURL, options: .atomic)
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        loadedCloneAudioSamples: [0.1, 0.2, 0.3],
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    await runtime.accept(
-        line: #"{"id":"req-relative-clone","op":"create_voice_profile_from_audio","profile_name":"ghost-copy","reference_audio_path":"reference.wav","vibe":"masc","transcript":"Provided transcript","cwd":"\#(callerWorkingDirectory.path)"}"#,
-    )
-
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == "req-relative-clone"
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "ghost-copy"
-        }
-    })
-}
-
-@Test func `create clone can infer transcript from reference audio`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-    try FileManager.default.createDirectory(at: storeRoot, withIntermediateDirectories: true)
-
-    let referenceAudioURL = storeRoot.appendingPathComponent("reference.wav")
-    try Data([0x01]).write(to: referenceAudioURL, options: .atomic)
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        loadedCloneAudioSamples: [0.4, 0.5, 0.6],
-        residentModelLoader: { _ in makeResidentModel() },
-        cloneTranscriptionModelLoader: {
-            makeCloneTranscriptionModel(transcript: "Inferred transcript")
-        },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let createID = await runtime.voices
-        .create(clone: "ghost-copy",
-                from: referenceAudioURL,
-                transcript: nil)
-        .id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == createID
-                && $0["ok"] as? Bool == true
-                && $0["profile_name"] as? String == "ghost-copy"
-        }
-    })
-    #expect(output.containsJSONObject {
-        $0["id"] as? String == createID
-            && $0["event"] as? String == "progress"
-            && $0["stage"] as? String == "loading_clone_transcription_model"
-    })
-    #expect(output.containsJSONObject {
-        $0["id"] as? String == createID
-            && $0["event"] as? String == "progress"
-            && $0["stage"] as? String == "transcribing_clone_audio"
-    })
-
-    let storedProfile = try makeProfileStore(rootURL: storeRoot).loadProfile(named: "ghost-copy")
-    #expect(storedProfile.manifest.sourceText == "Inferred transcript")
-}
-
-@Test func `create clone fails when transcript inference returns nothing`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-    try FileManager.default.createDirectory(at: storeRoot, withIntermediateDirectories: true)
-
-    let referenceAudioURL = storeRoot.appendingPathComponent("reference.wav")
-    try Data([0x01]).write(to: referenceAudioURL, options: .atomic)
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        loadedCloneAudioSamples: [0.7, 0.8, 0.9],
-        residentModelLoader: { _ in makeResidentModel() },
-        cloneTranscriptionModelLoader: {
-            makeCloneTranscriptionModel(transcript: "")
-        },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let createID = await runtime.voices
-        .create(clone: "ghost-copy",
-                from: referenceAudioURL,
-                transcript: nil)
-        .id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == createID
-                && $0["ok"] as? Bool == false
-                && $0["code"] as? String == "model_generation_failed"
-        }
-    })
-}
-
-@Test func `typed status and request streams expose worker output for library consumers`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let store = try makeProfileStore(rootURL: storeRoot)
-    _ = try store.createProfile(
-        profileName: "default-femme",
-        modelRepo: "test-model",
-        voiceDescription: "Warm and bright.",
-        sourceText: "Reference transcript",
-        sampleRate: 24000,
-        canonicalAudioData: Data([0x01, 0x02]),
-    )
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    let statuses = await runtime.statusEvents()
-    var statusIterator = statuses.makeAsyncIterator()
-
-    await runtime.start()
-
-    let firstStatus = await statusIterator.next()
-    let secondStatus = await statusIterator.next()
-    #expect(firstStatus == WorkerStatusEvent(stage: .warmingResidentModel, residentState: .warming, speechBackend: .qwen3))
-    #expect(secondStatus == WorkerStatusEvent(stage: .residentModelReady, residentState: .ready, speechBackend: .qwen3))
-
-    let handle = await runtime.submit(
-        .listProfiles(id: "req-stream"),
-    )
-    var iterator = handle.events.makeAsyncIterator()
-    let createdAt = try store.loadProfile(named: "default-femme").manifest.createdAt
-
-    var sawStarted = false
-    var sawCompleted = false
-
-    while let event = try await iterator.next() {
-        switch event {
-            case .queued:
-                continue
-
-            case let .started(started):
-                #expect(started == WorkerStartedEvent(id: "req-stream", op: "list_voice_profiles"))
-                sawStarted = true
-
-            case let .completed(response):
-                #expect(
-                    response == WorkerSuccessResponse(
-                        id: "req-stream",
-                        profiles: [
-                            ProfileSummary(
-                                profileName: "default-femme",
-                                vibe: .femme,
-                                createdAt: createdAt,
-                                voiceDescription: "Warm and bright.",
-                                sourceText: "Reference transcript",
-                                transcriptSource: nil,
-                                transcriptResolvedAt: nil,
-                                transcriptionModelRepo: nil,
-                            ),
-                        ],
-                    ),
-                )
-                sawCompleted = true
-
-            case .progress, .acknowledged:
-                Issue.record("The typed list-voice-profiles request stream emitted an unexpected event before completion: \(event)")
-        }
-    }
-
-    #expect(sawStarted)
-    #expect(sawCompleted)
-}
-
-@Test func `typed list profiles response includes transcript provenance summary`() async throws {
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let store = try makeProfileStore(rootURL: storeRoot)
-    let transcriptResolvedAt = Date(timeIntervalSince1970: 1_712_345_679)
-
-    _ = try store.createProfile(
-        profileName: "clone-profile",
-        vibe: .masc,
-        modelRepo: ModelFactory.importedCloneModelRepo,
-        voiceDescription: ModelFactory.importedCloneVoiceDescription,
-        sourceText: "Recovered transcript",
-        transcriptProvenance: TranscriptProvenance(
-            source: .inferred,
-            createdAt: transcriptResolvedAt,
-            transcriptionModelRepo: ModelFactory.cloneTranscriptionModelRepo,
-        ),
-        sampleRate: 24000,
-        canonicalAudioData: Data([0x52, 0x49, 0x46, 0x46]),
-    )
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: OutputRecorder(),
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-
-    let createdAt = try store.loadProfile(named: "clone-profile").manifest.createdAt
-    let handle = await runtime.submit(WorkerRequest.listProfiles(id: "req-provenance"))
-    var iterator = handle.events.makeAsyncIterator()
-
-    while let event = try await iterator.next() {
-        switch event {
-            case .queued, .started:
-                continue
-
-            case let .completed(response):
-                #expect(
-                    response == WorkerSuccessResponse(
-                        id: "req-provenance",
-                        profiles: [
-                            ProfileSummary(
-                                profileName: "clone-profile",
-                                vibe: .masc,
-                                createdAt: createdAt,
-                                voiceDescription: ModelFactory.importedCloneVoiceDescription,
-                                sourceText: "Recovered transcript",
-                                transcriptSource: .inferred,
-                                transcriptResolvedAt: transcriptResolvedAt,
-                                transcriptionModelRepo: ModelFactory.cloneTranscriptionModelRepo,
-                            ),
-                        ],
-                    ),
-                )
-                return
-
-            case .progress, .acknowledged:
-                Issue.record("The typed list-voice-profiles request stream emitted an unexpected event before completion: \(event)")
-        }
-    }
-
-    Issue.record("The typed list-voice-profiles request stream finished without emitting a completion event.")
-}
-
-@Test func `request observation returns nil and finished stream for unknown request ID`() async throws {
-    let runtime = try await makeRuntime(
-        output: OutputRecorder(),
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-
-    #expect(await runtime.request(id: "missing-request") == nil)
-
-    let updates = await runtime.updates(for: "missing-request")
-    var iterator = updates.makeAsyncIterator()
-    let first = try await iterator.next()
-    #expect(first == nil)
-}
-
-@Test func `request observation exposes replayable generation events for qwen requests`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let store = try makeProfileStore(rootURL: storeRoot)
-    _ = try store.createProfile(
-        profileName: "default-femme",
-        modelRepo: "test-model",
-        voiceDescription: "Warm and bright.",
-        sourceText: "Reference transcript",
-        sampleRate: 24000,
-        canonicalAudioData: Data([0x01, 0x02]),
-    )
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel(chunkCount: 2) },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let handle = await runtime.generate.audio(
-        text: "Hello from the generation event side channel.",
-        with: "default-femme",
-    )
-
-    let runtimeEvents = await runtime.generationEvents(for: handle.id)
-    var runtimeIterator = runtimeEvents.makeAsyncIterator()
-
-    if case let .token(token)? = try await runtimeIterator.next()?.event {
-        #expect(token == 101)
-    } else {
-        Issue.record("Expected the first replayed generation event to be a qwen token.")
-    }
-
-    if case let .info(info)? = try await runtimeIterator.next()?.event {
-        #expect(info.promptTokenCount == 12)
-        #expect(info.generationTokenCount == 8)
-        #expect(info.prefillTime == 0.12)
-        #expect(info.generateTime == 0.34)
-        #expect(info.tokensPerSecond == 56.7)
-        #expect(info.peakMemoryUsage == 1.23)
-    } else {
-        Issue.record("Expected the second replayed generation event to carry qwen generation info.")
-    }
-
-    if case let .audioChunk(sampleCount)? = try await runtimeIterator.next()?.event {
-        #expect(sampleCount == 2)
-    } else {
-        Issue.record("Expected the third replayed generation event to describe the first audio chunk.")
-    }
-
-    if case let .audioChunk(sampleCount)? = try await runtimeIterator.next()?.event {
-        #expect(sampleCount == 2)
-    } else {
-        Issue.record("Expected the fourth replayed generation event to describe the second audio chunk.")
-    }
-
-    #expect(try await runtimeIterator.next() == nil)
-
-    var handleIterator = handle.generationEvents.makeAsyncIterator()
-    if case let .token(token)? = try await handleIterator.next()?.event {
-        #expect(token == 101)
-    } else {
-        Issue.record("Expected the original RequestHandle generation stream to retain the qwen token event.")
-    }
-    if case let .info(info)? = try await handleIterator.next()?.event {
-        #expect(info.promptTokenCount == 12)
-    } else {
-        Issue.record("Expected the original RequestHandle generation stream to retain the qwen info event.")
-    }
-    if case let .audioChunk(sampleCount)? = try await handleIterator.next()?.event {
-        #expect(sampleCount == 2)
-    } else {
-        Issue.record("Expected the original RequestHandle generation stream to retain the first audio chunk event.")
-    }
-    if case let .audioChunk(sampleCount)? = try await handleIterator.next()?.event {
-        #expect(sampleCount == 2)
-    } else {
-        Issue.record("Expected the original RequestHandle generation stream to retain the second audio chunk event.")
-    }
-    #expect(try await handleIterator.next() == nil)
-}
-
-@Test func `request observation replays queued state and fans out to multiple subscribers`() async throws {
-    let output = OutputRecorder()
-    let preloadGate = AsyncGate()
-    let runtime = try await makeRuntime(
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in
-            await preloadGate.wait()
-            return makeResidentModel()
-        },
-    )
-
-    await runtime.start()
-
-    let handle = await runtime.submit(.listProfiles(id: "req-late"))
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == "req-late"
-                && $0["event"] as? String == "queued"
-                && $0["reason"] as? String == "waiting_for_resident_model"
-        }
-    })
-
-    let snapshot = await runtime.request(id: "req-late")
-    #expect(snapshot?.id == "req-late")
-    #expect(snapshot?.operation == "list_voice_profiles")
-    #expect(snapshot?.sequence == 1)
-    if let snapshot {
-        switch snapshot.state {
-            case let .queued(queued):
-                #expect(queued == WorkerQueuedEvent(id: "req-late", reason: .waitingForResidentModel, queuePosition: 1))
-            default:
-                Issue.record("Expected queued state in the late-attach snapshot, got \(snapshot.state)")
-        }
-        #expect(snapshot.acceptedAt <= snapshot.lastUpdatedAt)
-    }
-
-    let updatesA = await runtime.updates(for: "req-late")
-    let updatesB = await runtime.updates(for: "req-late")
-    var iteratorA = updatesA.makeAsyncIterator()
-    var iteratorB = updatesB.makeAsyncIterator()
-
-    let replayA = try await iteratorA.next()
-    let replayB = try await iteratorB.next()
-    #expect(replayA?.sequence == 1)
-    #expect(replayB?.sequence == 1)
-    if case let .queued(queuedA)? = replayA?.state {
-        #expect(queuedA.reason == .waitingForResidentModel)
-    } else {
-        Issue.record("Expected subscriber A to replay a queued update first.")
-    }
-    if case let .queued(queuedB)? = replayB?.state {
-        #expect(queuedB.reason == .waitingForResidentModel)
-    } else {
-        Issue.record("Expected subscriber B to replay a queued update first.")
-    }
-
-    await preloadGate.open()
-
-    let startedA = try await iteratorA.next()
-    let startedB = try await iteratorB.next()
-    #expect(startedA?.sequence == 2)
-    #expect(startedB?.sequence == 2)
-    if case let .started(eventA)? = startedA?.state {
-        #expect(eventA == WorkerStartedEvent(id: "req-late", op: "list_voice_profiles"))
-    } else {
-        Issue.record("Expected subscriber A to receive a started update second.")
-    }
-    if case let .started(eventB)? = startedB?.state {
-        #expect(eventB == WorkerStartedEvent(id: "req-late", op: "list_voice_profiles"))
-    } else {
-        Issue.record("Expected subscriber B to receive a started update second.")
-    }
-
-    let completedA = try await iteratorA.next()
-    let completedB = try await iteratorB.next()
-    #expect(completedA?.sequence == 3)
-    #expect(completedB?.sequence == 3)
-    if case let .completed(successA)? = completedA?.state {
-        #expect(successA.id == "req-late")
-    } else {
-        Issue.record("Expected subscriber A to receive a completed update third.")
-    }
-    if case let .completed(successB)? = completedB?.state {
-        #expect(successB.id == "req-late")
-    } else {
-        Issue.record("Expected subscriber B to receive a completed update third.")
-    }
-
-    #expect(try await iteratorA.next() == nil)
-    #expect(try await iteratorB.next() == nil)
-
-    var handleIterator = handle.events.makeAsyncIterator()
-    let handleQueued = try await handleIterator.next()
-    let handleStarted = try await handleIterator.next()
-    let handleCompleted = try await handleIterator.next()
-    if case let .queued(queued)? = handleQueued {
-        #expect(queued == WorkerQueuedEvent(id: "req-late", reason: .waitingForResidentModel, queuePosition: 1))
-    } else {
-        Issue.record("Expected the original handle stream to retain the queued event history.")
-    }
-    if case let .started(started)? = handleStarted {
-        #expect(started == WorkerStartedEvent(id: "req-late", op: "list_voice_profiles"))
-    } else {
-        Issue.record("Expected the original handle stream to retain the started event history.")
-    }
-    if case let .completed(success)? = handleCompleted {
-        #expect(success.id == "req-late")
-    } else {
-        Issue.record("Expected the original handle stream to retain the completed event history.")
-    }
-    #expect(try await handleIterator.next() == nil)
-
-    let completedSnapshot = await runtime.request(id: "req-late")
-    #expect(completedSnapshot?.sequence == 3)
-    if case let .completed(success)? = completedSnapshot?.state {
-        #expect(success.id == "req-late")
-    } else {
-        Issue.record("Expected the retained request snapshot to stay completed after terminal success.")
-    }
-}
-
-@Test func `request observation reports cancellation as data while handle stream still throws`() async throws {
-    let output = OutputRecorder()
-    let profileGate = AsyncGate()
-
-    let runtime = try await makeRuntime(
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-        profileModelLoader: {
-            makeProfileModel {
-                await profileGate.wait()
-            }
-        },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    _ = await runtime.voices.create(design: "bright-guide",
-                                    from: "Hello there",
-                                    vibe: .femme,
-                                    voice: "Warm and bright")
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "started"
-                && $0["op"] as? String == "create_voice_profile_from_description"
-        }
-    })
-
-    let queuedHandle = await runtime.submit(.listProfiles(id: "req-cancelled"))
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == "req-cancelled"
-                && $0["event"] as? String == "queued"
-                && $0["reason"] as? String == "waiting_for_active_request"
-        }
-    })
-
-    let updates = await runtime.updates(for: "req-cancelled")
-    var updatesIterator = updates.makeAsyncIterator()
-    if case let .queued(queued)? = try await updatesIterator.next()?.state {
-        #expect(queued == WorkerQueuedEvent(id: "req-cancelled", reason: .waitingForActiveRequest, queuePosition: 1))
-    } else {
-        Issue.record("Expected the reconnecting observer to replay the queued state first.")
-    }
-
-    let cancelID = await runtime.player.cancelRequest("req-cancelled").id
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == cancelID
-                && $0["ok"] as? Bool == true
-                && $0["cancelled_request_id"] as? String == "req-cancelled"
-        }
-    })
-
-    let cancelledUpdate = try await updatesIterator.next()
-    if case let .cancelled(failure)? = cancelledUpdate?.state {
-        #expect(failure.id == "req-cancelled")
-        #expect(failure.code == .requestCancelled)
-    } else {
-        Issue.record("Expected the reconnecting observer to receive cancellation as data.")
-    }
-    #expect(try await updatesIterator.next() == nil)
-
-    let cancelledSnapshot = await runtime.request(id: "req-cancelled")
-    if case let .cancelled(failure)? = cancelledSnapshot?.state {
-        #expect(failure.code == .requestCancelled)
-    } else {
-        Issue.record("Expected the retained request snapshot to stay cancelled after terminal failure.")
-    }
-
-    var handleIterator = queuedHandle.events.makeAsyncIterator()
-    do {
-        while let _ = try await handleIterator.next() {}
-        Issue.record("The original RequestHandle stream should still throw on cancellation for compatibility.")
-    } catch let error as WorkerError {
-        #expect(error.code == .requestCancelled)
-    }
-
-    await profileGate.open()
-}
-
-@Test func `speak live uses stable resident generation parameters`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let store = try makeProfileStore(rootURL: storeRoot)
-    _ = try store.createProfile(
-        profileName: "default-femme",
-        modelRepo: "test-model",
-        voiceDescription: "Warm and bright.",
-        sourceText: "Reference transcript",
-        sampleRate: 24000,
-        canonicalAudioData: Data([0x01, 0x02]),
-    )
-
-    let recorder = ResidentModelRecorder()
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel(recorder: recorder) },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let requestID = await runtime.generate
-        .speech(
-            text: "Hello there, galew.",
-            with: "default-femme",
-        )
-        .id
-
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == requestID
-                && $0["event"] as? String == "progress"
-                && $0["stage"] as? String == "preroll_ready"
-        }
-    })
-
-    #expect(recorder.lastText == "Hello there, gale wumbo.")
-    #expect(recorder.lastGenerationParameters?.maxTokens == 56)
-    #expect(recorder.lastGenerationParameters?.temperature == 0.9)
-    #expect(recorder.lastGenerationParameters?.topP == 1.0)
-    #expect(recorder.lastGenerationParameters?.repetitionPenalty == 1.05)
-}
-
-@Test func `late status subscribers receive current ready snapshot`() async throws {
-    let output = OutputRecorder()
-    let runtime = try await makeRuntime(
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let statuses = await runtime.statusEvents()
-    var iterator = statuses.makeAsyncIterator()
-
-    #expect(await iterator.next() == WorkerStatusEvent(stage: .residentModelReady, residentState: .ready, speechBackend: .qwen3))
-}
-
-@Test func `dropping status subscription does not retain runtime`() async throws {
-    let output = OutputRecorder()
-    let weakRuntime = WeakRuntimeBox()
-
-    var runtime: WorkerRuntime? = try await makeRuntime(
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-    weakRuntime.value = runtime
-
-    var statuses: AsyncStream<SpeakSwiftly.StatusEvent>? = await runtime?.statusEvents()
-    _ = statuses?.makeAsyncIterator()
-
-    statuses = nil
-    runtime = nil
-
-    #expect(await waitUntil { weakRuntime.value == nil })
-}
-
-@Test func `start is idempotent for library consumers`() async throws {
-    actor LoadCounter {
-        private(set) var count = 0
-
-        func increment() {
-            count += 1
-        }
-
-        func value() -> Int {
-            count
-        }
-    }
-
-    let output = OutputRecorder()
-    let loadCounter = LoadCounter()
-    let runtime = try await makeRuntime(
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in
-            await loadCounter.increment()
-            return makeResidentModel()
-        },
-    )
-
-    let statuses = await runtime.statusEvents()
-    var iterator = statuses.makeAsyncIterator()
-
-    await runtime.start()
-    await runtime.start()
-
-    let firstStatus = await iterator.next()
-    let secondStatus = await iterator.next()
-
-    #expect(firstStatus == WorkerStatusEvent(stage: .warmingResidentModel, residentState: .warming, speechBackend: .qwen3))
-    #expect(secondStatus == WorkerStatusEvent(stage: .residentModelReady, residentState: .ready, speechBackend: .qwen3))
-    #expect(await loadCounter.value() == 1)
-    #expect(
-        output.countJSONObjects {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "warming_resident_model"
-        } == 1,
-    )
-    #expect(
-        output.countJSONObjects {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        } == 1,
-    )
-}
-
-@Test func `typed request stream keeps background acknowledgement and later completion separate`() async throws {
-    let output = OutputRecorder()
-    let playbackDrain = AsyncGate()
-    let playback = PlaybackSpy(behavior: .gate(playbackDrain))
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let store = try makeProfileStore(rootURL: storeRoot)
-    _ = try store.createProfile(
-        profileName: "default-femme",
-        modelRepo: "test-model",
-        voiceDescription: "Warm and bright.",
-        sourceText: "Reference transcript",
-        sampleRate: 24000,
-        canonicalAudioData: Data([0x01, 0x02]),
-    )
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: playback,
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    let activeHandle = await runtime.generate.speech(text: "Hello there", with: "default-femme")
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["id"] as? String == activeHandle.id
-                && $0["event"] as? String == "progress"
-                && $0["stage"] as? String == "preroll_ready"
-        }
-    })
-
-    let handle = await runtime.submit(
-        .queueSpeech(
-            id: "req-stream-bg",
-            text: "Hi there",
-            profileName: "default-femme",
-            textProfileName: nil,
-            jobType: .live,
-            textContext: nil,
-            sourceFormat: nil,
-        ),
-    )
-    var iterator = handle.events.makeAsyncIterator()
-
-    let acknowledged = try await iterator.next()
-    #expect(acknowledged == .acknowledged(WorkerSuccessResponse(id: "req-stream-bg")))
-
-    await playbackDrain.open()
-
-    var sawCompletion = false
-    while let event = try await iterator.next() {
-        if case .completed(WorkerSuccessResponse(id: "req-stream-bg", profileName: nil, profilePath: nil, profiles: nil)) = event {
-            sawCompletion = true
-            break
-        }
-    }
-
-    #expect(sawCompletion)
-}
-
-@Test func `list profiles skips corrupt entries and still returns healthy profiles`() async throws {
-    let output = OutputRecorder()
-    let storeRoot = makeTempDirectoryURL()
-    defer { try? FileManager.default.removeItem(at: storeRoot) }
-
-    let store = try makeProfileStore(rootURL: storeRoot)
-    _ = try store.createProfile(
-        profileName: "healthy",
-        modelRepo: "test-model",
-        voiceDescription: "Warm and bright.",
-        sourceText: "Healthy transcript",
-        sampleRate: 24000,
-        canonicalAudioData: Data([0x01]),
-    )
-
-    let brokenDirectory = store.profileDirectoryURL(for: "broken")
-    try FileManager.default.createDirectory(at: brokenDirectory, withIntermediateDirectories: false)
-    try Data("not-json".utf8).write(to: store.manifestURL(for: brokenDirectory))
-
-    let runtime = try await makeRuntime(
-        rootURL: storeRoot,
-        output: output,
-        playback: PlaybackSpy(),
-        residentModelLoader: { _ in makeResidentModel() },
-    )
-
-    await runtime.start()
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            $0["event"] as? String == "worker_status"
-                && $0["stage"] as? String == "resident_model_ready"
-        }
-    })
-
-    await runtime.accept(line: #"{"id":"req-1","op":"list_voice_profiles"}"#)
-
-    #expect(await waitUntil {
-        output.containsJSONObject {
-            guard
-                $0["id"] as? String == "req-1",
-                $0["ok"] as? Bool == true,
-                let profiles = $0["profiles"] as? [[String: Any]]
-            else {
-                return false
-            }
-
-            return profiles.count == 1
-                && profiles[0]["profile_name"] as? String == "healthy"
-        }
-    })
-}
-
 @Test func `waiting speak live runs before waiting profile management after active work finishes`() async throws {
     let output = OutputRecorder()
     let playback = PlaybackSpy()
@@ -2388,9 +998,3 @@ private actor BackendLoadRecorder {
     #expect(startedOps == ["req-1:create_voice_profile_from_description", "req-2:generate_speech", "req-3:list_voice_profiles"])
 }
 
-private func rawTestAudioData(for samples: [Float]) -> Data {
-    let bytes = samples.map(\.bitPattern).flatMap { value in
-        withUnsafeBytes(of: value.littleEndian, Array.init)
-    }
-    return Data(bytes)
-}
