+++ Tests/SpeakSwiftlyTests/Storage/GeneratedFileStoreTests.swift
